Compatiblity with custom pokeexpansion
There was a problem with the learnset array on a rom with expanded mons not equal to 439, 546, 702, 774 and smaller then 775.  This edit changes the EMPTYSLOT_MOVESET assigner in a way that it doesn't assign an EMPTYSLOT_MOVESET to a non-existent pokemon. Now the array won't be exceeded. This also makes it work with PGE